Ajustado estilos da tela principal.
Ajustado estilo da tela principal;
Melhorado mensagem de erro e sucesso ao realizar cadastro e login.
